Define login_form contents in fewer places
Move to a template in index.html for defining the contents of
login_form when not actually logged in.

Previously, we had several places where the HTML contents of the
login form were stored, some different than others.

This unifies, while still leaving one "duplicate" set of contents in
place in the actual index.html (not the template).

From what I can see currently, this baseline is always replaced with
the template, and the baseline could be removed in the future.

Signed-off-by: Jeff Godin <jgodin@tadl.org>